# docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      error:
        type: string
    type: object
...
  models.GearCategory:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      category_top_category_id:
        type: integer
    type: object
...
  models.GearTopCategory:
    properties:
      top_category_id:
        type: integer
      top_category_name:
        type: string
    type: object
...
  models.Manufacture:
    properties:
      manufacture_id:
        type: integer
      manufacture_name:
        type: string
    type: object
  models.ResponsePayload:
    properties:
      current_page:
        type: integer
      item_limit:
        type: integer
      items: {}
      next_page:
        type: string
      prev_page:
        type: string
      total_item_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
  models.UserContainerNoID:
    properties:
      user_container_id:
        type: integer
      user_gear_registration_id:
        type: integer
    type: object
...
  models.UserGearLinkNoID:
    properties:
      usergear_gear_id:
        type: integer
      usergear_user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@sea-shell.no
    name: API Support
  description: This is the API of GoGear
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoGear API
  version: "1.0"
paths:
  /api/v1/category/{category}/delete:
    delete:
      consumes:
      - application/json
      description: Delete category with corresponding ID value. Requires a JWT issued
        with the admin audience.
      parameters:
      - description: Unique ID of category you want to update
        in: path
        name: category
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete category with ID
      tags:
      - Category
  /api/v1/category/{category}/update:
    post:
      consumes:
      - application/json
      description: Update category identified by ID. Requires a JWT issued with the
        admin audience.
      parameters:
      - description: Unique ID of category you want to update
        in: path
        name: categoryID
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GearCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update category with ID
      tags:
      - Category
  /api/v1/category/{categoryID}/get:
    get:
      consumes:
      - application/json
      description: Get category spessific to ID
      parameters:
      - description: Unique ID of category you want to get
        in: path
        name: categoryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/models.GearCategory'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get category with ID
      tags:
      - Category
  /api/v1/category/insert:
    put:
      consumes:
      - application/json
      description: Insert new category with corresponding values. Requires a JWT issued
        with the admin audience.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GearCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.GearCategory'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Insert new category
      tags:
      - Category
  /api/v1/category/list:
    get:
      consumes:
      - application/json
      description: Get a list of category items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - collectionFormat: multi
        description: Top category
        in: query
        items:
          type: integer
        name: category
        type: array
      - collectionFormat: multi
        description: Top gear category
        in: query
        items:
          type: integer
        name: topCategory
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.GearCategory'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Category
  /api/v1/container/{container}/delete:
    delete:
      consumes:
      - application/json
      description: Delete user container content with corresponding ID value
      parameters:
      - description: Unique ID of userGear you want to update
        in: path
        name: container
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete user container content with ID
      tags:
      - User container
  /api/v1/container/{container}/list:
    get:
      consumes:
      - application/json
      description: List all items inside a container registered to a user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Unique ID of userGear you want to update
        in: path
        name: container
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.FullGear'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List all items inside a container
      tags:
      - User container
  /api/v1/container/insert:
    put:
      consumes:
      - application/json
      description: Insert user registered gear with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserContainerNoID'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
      security:
      - BearerAuth: []
      summary: Insert user registered gear
      tags:
      - User container
  /api/v1/gear/{gear}/delete:
    delete:
      consumes:
      - application/json
      description: Delete gear with corresponding ID value
      parameters:
      - description: Unique ID of gear you want to delete
        in: path
        name: gear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete gear with ID
      tags:
      - Gear
  /api/v1/gear/{gear}/get:
    get:
      consumes:
      - application/json
      description: Get gear spessific to ID
      parameters:
      - description: Unique ID of Gear you want to get
        in: path
        name: gear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/models.FullGear'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get gear with ID
      tags:
      - Gear
  /api/v1/gear/{gear}/update:
    post:
      consumes:
      - application/json
      description: Update gear identified by ID
      parameters:
      - description: Unique ID of Gear you want to get
        in: path
        name: gear
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Gear'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update gear with ID
      tags:
      - Gear
  /api/v1/gear/insert:
    put:
      consumes:
      - application/json
      description: Insert new gear with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Gear'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Gear'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Insert new gear
      tags:
      - Gear
  /api/v1/gear/list:
    get:
      consumes:
      - application/json
      description: Get a list of gear items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Gear category
        in: query
        name: category
        type: string
      - description: Top gear category
        in: query
        name: topCategory
        type: string
      - description: Gear manufacturer
        in: query
        name: manufacturer
        type: string
      - collectionFormat: multi
        description: string collection
        in: query
        items:
          type: string
        name: collection
        type: array
      - default: false
        description: string collection
        in: query
        name: container
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.GearListItem'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List gear
      tags:
      - Gear
  /api/v1/gear/search:
    get:
      consumes:
      - application/json
      description: Get a list of gear items based on search patterns
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: String to search for
        in: query
        name: searchString
        required: true
        type: string
      - description: 'Type of searhc method. valid choices are: startswith, contains,
          endswith'
        in: query
        name: searchType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.GearListItem'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Search for gear
      tags:
      - Gear
  /api/v1/manufacture/{manufacture}/delete:
    delete:
      consumes:
      - application/json
      description: Delete manufacture with corresponding ID value
      parameters:
      - description: Unique ID of manufacture you want to update
        in: path
        name: manufacture
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete manufacture with ID
      tags:
      - Manufacture
  /api/v1/manufacture/{manufacture}/get:
    get:
      consumes:
      - application/json
      description: Get manufacture spessific to ID
      parameters:
      - description: Unique ID of manufacture you want to get
        in: path
        name: manufacture
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/models.Manufacture'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get manufacture by ID
      tags:
      - Manufacture
  /api/v1/manufacture/{manufacture}/update:
    post:
      consumes:
      - application/json
      description: Update manufacture identified by ID
      parameters:
      - description: Unique ID of manufacture you want to update
        in: path
        name: manufacture
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Manufacture'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update manufacture with ID
      tags:
      - Manufacture
  /api/v1/manufacture/insert:
    put:
      consumes:
      - application/json
      description: Insert new manufacture with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Manufacture'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Manufacture'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Insert new manufacture
      tags:
      - Manufacture
  /api/v1/manufacture/list:
    get:
      consumes:
      - application/json
      description: Get a list of manufacturers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: search by manufacturename (this is case insensitive and wildcard)
        in: query
        name: manufacture
        type: string
      - description: search by manufactures full name (this is case insensitive and
          wildcard)
        in: query
        name: manufacturename
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.Manufacture'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List manufacture
      tags:
      - Manufacture
  /api/v1/topCategory/{topCategory}/delete:
    delete:
      consumes:
      - application/json
      description: Delete topCategory with corresponding ID value
      parameters:
      - description: Unique ID of topCategory you want to update
        in: path
        name: topCategory
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete topCategory with ID
      tags:
      - Top Category
  /api/v1/topCategory/{topCategory}/get:
    get:
      consumes:
      - application/json
      description: Get top category spessific to ID
      parameters:
      - description: Unique ID of top category you want to get
        in: path
        name: topCategoryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GearTopCategory'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get top category with ID
      tags:
      - Top Category
  /api/v1/topCategory/{topCategory}/update:
    post:
      consumes:
      - application/json
      description: Update top category identified by ID
      parameters:
      - description: Unique ID of top category you want to update
        in: path
        name: topCategoryID
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GearTopCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update top category with ID
      tags:
      - Top Category
  /api/v1/topCategory/insert:
    put:
      consumes:
      - application/json
      description: Insert new top category with corresponding values
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GearTopCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Insert new top category
      tags:
      - Top Category
  /api/v1/topCategory/list:
    get:
      consumes:
      - application/json
      description: Get a list of top category items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - collectionFormat: multi
        description: top categories
        in: query
        items:
          type: integer
        name: topCategory
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.GearTopCategory'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List top categories
      tags:
      - Top Category
  /api/v1/usergear/{user}/list:
    get:
      consumes:
      - application/json
      description: Get a list a users gear
      parameters:
      - description: Unique ID of user you want to get the Gear of
        in: path
        name: user
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - collectionFormat: multi
        description: top categories
        in: query
        items:
          type: integer
        name: topCategory
        type: array
      - collectionFormat: multi
        description: sub categories
        in: query
        items:
          type: integer
        name: category
        type: array
      - collectionFormat: multi
        description: manufacturers
        in: query
        items:
          type: integer
        name: manufacture
        type: array
      - default: all
        description: show container gear only. valid values are true, false, all
        in: query
        name: container
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.UserGear'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: List users gear
      tags:
      - User gear
  /api/v1/usergear/insert:
    put:
      consumes:
      - application/json
      description: Insert user registered gear with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserGearLinkNoID'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
      security:
      - BearerAuth: []
      summary: Insert user registered gear
      tags:
      - User gear
  /api/v1/usergear/registration/{usergear}/delete:
    delete:
      consumes:
      - application/json
      description: Delete userGear with corresponding ID value
      parameters:
      - description: Unique ID of userGear you want to update
        in: path
        name: userGear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete userGear with ID
      tags:
      - User gear
  /api/v1/usergear/registration/{usergear}/get:
    get:
      consumes:
      - application/json
      description: Get user registeredgear spessific to ID
      parameters:
      - description: Unique ID of user registered gear you want to get
        in: path
        name: usergear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/models.UserGear'
      security:
      - BearerAuth: []
      summary: Get user registered gear with ID
      tags:
      - User gear
  /api/v1/usergear/registration/{usergear}/update:
    post:
      consumes:
      - application/json
      description: Update user registered gear identified by ID
      parameters:
      - description: Unique ID of user registered gear you want to get
        in: path
        name: usergear
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserGearLink'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
      security:
      - BearerAuth: []
      summary: Update user registered gear with ID
      tags:
      - User gear
...
securityDefinitions:
  BearerAuth:
    description: Include a server-issued JWT as `Bearer <token>`. Endpoints may require
      either the client or admin audience.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"